To avoid an error on DALI, making a new module for dataset.
