+++ examples/imagenet/dataset_util.py
@@ -1,0 +1,45 @@
+import random
+
+import chainer
+from chainer import dataset
+
+
+class PreprocessedDataset(chainer.dataset.DatasetMixin):
+
+    def __init__(self, path, root, mean, crop_size, random=True):
+        self.base = chainer.datasets.LabeledImageDataset(path, root)
+        self.mean = mean.astype(chainer.get_dtype())
+        self.crop_size = crop_size
+        self.random = random
+
+    def __len__(self):
+        return len(self.base)
+
+    def get_example(self, i):
+        # It reads the i-th image/label pair and return a preprocessed image.
+        # It applies following preprocesses:
+        #     - Cropping (random or center rectangular)
+        #     - Random flip
+        #     - Scaling to [0, 1] value
+        crop_size = self.crop_size
+
+        image, label = self.base[i]
+        _, h, w = image.shape
+
+        if self.random:
+            # Randomly crop a region and flip the image
+            top = random.randint(0, h - crop_size - 1)
+            left = random.randint(0, w - crop_size - 1)
+            if random.randint(0, 1):
+                image = image[:, :, ::-1]
+        else:
+            # Crop the center
+            top = (h - crop_size) // 2
+            left = (w - crop_size) // 2
+        bottom = top + crop_size
+        right = left + crop_size
+
+        image = image[:, top:bottom, left:right]
+        image -= self.mean[:, top:bottom, left:right]
+        image *= (1.0 / 255.0)  # Scale to [0, 1]
+        return image, label
